requirements.txt
autogluon
matplotlib
numpy
pandas
python-dotenv
Requests
streamlit
